Crate ic_crypto_internal_tls

Library crate that provides low-level functionality needed to establish TLS connections.

In particular, the crate provides functionality to

generate TLS key material and wrap the public part in an X.509 certificate

create an OpenSSL TLS acceptor

create an OpenSSL TLS connector

Modules

keygen
Functionality to generate key material to be used for TLS connections.

Structs

CreateTlsAcceptorError
A TLS acceptor couldn't be created.

CreateTlsConnectorError
A TLS connector couldn't be created.

Functions

tls_acceptor
Builds a TLS acceptor to establish TLS connections on the server side.

tls_connector
Builds a TLS connector to establish TLS connections on the client side.