Function ic_crypto_internal_tls::keygen::generate_tls_key_pair

pub fn generate_tls_key_pair(
    common_name: &str, 
    serial: [u8; 19], 
    not_after: &Asn1Time
) -> (X509, PKey<Private>)

Generate a key pair and return the certificate and private key.

Note that the certificate serial number must be at most 20 octets according to https://tools.ietf.org/html/rfc5280 Section 4.1.2.2. The 19 bytes serial number argument is interpreted as an unsigned integer and thus fits in 20 bytes, encoded as a signed ASN1 integer.